reorganized files and folders (model + utils)
