+++ edi_system/bin/model/electronic_device.dart
