+++ edi_system/bin/model/laptop.dart
