+++ edi_system/bin/model/smartphone.dart
